Document licenses, when possible.
Some code, particularly the BF interpreter, has no license because it
was not written to be distinct from the RPython/PyPy ecosystem. If I
ever find the Haskell implementation, then it will be in a similar
situation. For now, this is all fine; it's all Free Software, just
lacking the documentation required for corporate exploitation.