Use setup hooks to prepare build directories.
This was fairly frustrating. I had to dig through several examples of
setup hooks in nixpkgs, along with repeated lookups in nixpkgs
documentation, just to be relatively confident in which setup-hook
approach to use. The hooks themselves are simple and perform the same
patching steps that were used ad-hoc in derivations.