Add Pixie.
Also do a bunch of required fixups. Pixie doesn't yet build, but we are
starting to get some patterns about broken packages, since they are
broken in similar ways due to bitrot.

Topaz is still broken, and now marked as such. There is now a flag to
allow building with PyPy's full source tree, for Topaz or others which
need it.