ci: Pick out some good targets to build.
Any target will include pypy2Minimal, which is what *really* matters for
both the magic cache and Cachix. So, let's be sure to include some
targets worth testing. DIVSPL has a checkPhase. PyGirl binds SDL, which
is a fairly non-trivial dependency. PyPy is necessary to continue doing
this whole shebang.