bf: Discard all initial loops.
This removes comments from the top of many programs. The structure of
peep() prevented this from happening at the top level, so it must be
done explicitly. It could have been done implicitly by starting with a
zero op and peepholing on top of that (since loops after zero are also
discarded) but that zero would itself have to be removed later and it's
just not worth the hassle.