bf: Use tagless-final encoding for Brainfuck.
Using the algebraic presentation of Brainfuck semantics, we're able to
do code generation which is ultimately simpler, shorter, and faster.

I did have to pull out a few tricks to get the JIT back up to parity.
Because looping and sequencing are now decoupled, extra immutability
annotations are required as well as an unrolling annotation. But the end
result is great, showing an overall speedup on both mandel.b and
bench.b despite not matching all of their idioms.