bf: Fix handling of initial comments.
That's a pretty bad parser fuckup. So why did I want to do this instead
of a semantic approach? Well, the optimizer needs to know a special
starting condition that doesn't hold in general: the starting cell is
zero. But the only guaranteed zeroing operations are loops, so we'd have
to prepend a loop in order to optimize away a loop, and we'd always have
a three-character penalty from that prepended loop.

Other implementations could choose differently, but we want the -o flag
to print out a canonicalized version of the input so that bbgauge.info
and similar projects aren't biased.