pysom: Split to pysom-ast, pysom-bc.
The old name will point to the bytecode interpreter, which appears to be
the one preferred upstream. Both are checked and there's no regression.

This maker pattern seems like the best way to add this sort of
build-time flavoring, and it worked so cleanly here that I'm imagining
that it will work just as well on, say, Pydrofoil.