Package SAIL/ISLA tooling.
SAIL is a language for specifying CPUs. It has two toolchains, alike in
dignity and jank. We need the OCaml toolchain for Pydrofoil. ISLA is an
IR for SAIL, I think?

This commit doesn't actually include the three models that Pydrofoil
currently supports; those have to be built separately, and at this point
of the night I'd like to go to bed.