pygirl: Don't warn about new name.
Apparently the flake style is *not* to give warnings to users; a flake
ought to always build every output, which means that a broken output is
not communicating any extra information to users; it'd break either way.

My retort is simple: `throw` lets *me*, the person making the change,
send a short message to the user. That's literally a communication of
extra information.

But whatever. This should fix CI.